save_memory keyword added. For now, it just works in pure python mode.
A keyword "save_memory" is added to the train method. In this mode,
db is used as the data format, and fingerprints/neighborlists are
read/write "on the fly" from/on disk without being saved in memory. While the memory usage is not expected to increase too much with the number of
images, read/write time is expected to increase.

Autotests are also now extended with save_memory option.

A couple of points:
1. save_memory is True by default in the current version. But will be
switched to False, after a couple of daily test runs.
2. This mode will be extended to Fortran.